+++ pdbeccdutils/core/depictions.py
@@ -33,7 +33,7 @@ from pdbeccdutils.core.models import DepictionResult, DepictionSource
 from pdbeccdutils.helpers.mol_tools import fix_conformer
 from pdbeccdutils.utils import config
 from rdkit import Chem, Geometry
-from rdkit.Chem import AllChem, rdCoordGen
+from rdkit.Chem import AllChem, rdCoordGen, rdDepictor
 from scipy.spatial import KDTree
 
 
@@ -101,6 +101,7 @@ class DepictionManager:
 
         templateMol = Chem.RWMol(temp_mol).GetMol()
         pubchemMol = Chem.RWMol(temp_mol).GetMol()
+        rdkitMolCd = Chem.RWMol(temp_mol).GetMol()
         rdkitMol = Chem.RWMol(temp_mol).GetMol()
         results = []
 
@@ -110,11 +111,14 @@ class DepictionManager:
             else None
         )
         template_res = self._get_2D_by_template(templateMol) if self.templates else []
+        rdkit_res_cd = self._get_2D_by_rdkit_coordgen(rdkitMolCd)
         rdkit_res = self._get_2D_by_rdkit(rdkitMol)
 
-        if pubchem_res is not None:
+        if (pubchem_res is not None) and (pubchem_res.source.name == 'PubChem'):
             results.append(pubchem_res)
-        if rdkit_res is not None:
+        if rdkit_res_cd.source.name == 'RDKit':
+            results.append(rdkit_res_cd)
+        if rdkit_res.source.name == 'RDKit':
             results.append(rdkit_res)
 
         results = results + template_res
@@ -129,7 +133,7 @@ class DepictionManager:
 
         return DepictionResult(
             source=DepictionSource.Failed,
-            template_name="",
+            template_name=None,
             mol=None,
             score=float("inf"),
         )
@@ -180,6 +184,30 @@ class DepictionManager:
 
         return mol
 
+    def _get_2D_by_rdkit_coordgen(self, mol):
+        """
+        Get depiction done using solely the default RDKit coordgen functionality.
+
+        Args:
+            mol (rdkit.Chem.rdchem.Mol): Mol to be depicted
+
+        Returns:
+            DepictionResult: Depiction with some usefull metadata
+        """
+        try:
+            rdCoordGen.AddCoords(mol, self.coordgen_params)
+            flaws = DepictionValidator(mol).depiction_score()
+            return DepictionResult(
+                source=DepictionSource.RDKit, template_name=None, mol=mol, score=flaws
+            )
+        except Exception:
+            return DepictionResult(
+                source=DepictionSource.Failed,
+                template_name=None,
+                mol=None,
+                score=float("inf"),
+            )
+
     def _get_2D_by_rdkit(self, mol):
         """
         Get depiction done using solely the default RDKit functionality.
@@ -191,7 +219,8 @@ class DepictionManager:
             DepictionResult: Depiction with some usefull metadata
         """
         try:
-            rdCoordGen.AddCoords(mol, self.coordgen_params)
+            rdDepictor.SetPreferCoordGen(False)
+            rdDepictor.Compute2DCoords(mol, useRingTemplates=True)
             flaws = DepictionValidator(mol).depiction_score()
             return DepictionResult(
                 source=DepictionSource.RDKit, template_name=None, mol=mol, score=flaws